+++ tests/unittests/metric/product/unittest_metric_tests.py
@@ -55,9 +55,8 @@ class FakeSonar(object):
 
 class FakeSubject(object):
     """ Provide for a fake subject. """
-    def __init__(self, sonar=None, integration_tests=False):
-        self.__integration_tests = domain.Product(domain.Project(), metric_source_ids={sonar: 'some:fake:id'}) \
-            if integration_tests else None
+    def __init__(self, integration_tests=False):
+        self.__has_integration_tests = integration_tests
 
     @staticmethod
     def name():
@@ -69,9 +68,9 @@ class FakeSubject(object):
         """ Return the metric source id for Sonar. """
         return 'some:fake:id'
 
-    def integration_tests(self):
-        """ Return the integration tests of the subject. """
-        return self.__integration_tests
+    def has_integration_tests(self):
+        """ Return whether the subject has integration tests. """
+        return self.__has_integration_tests
 
 
 class CommonUnittestMetricTestsMixin(object):
@@ -85,7 +84,7 @@ class CommonUnittestMetricTestsMixin(object):
         """ Set up the fixture for the unit tests. """
         self.__sonar = FakeSonar(line_coverage=self.expected_value, branch_coverage=self.expected_value)
         project = domain.Project(metric_sources={metric_source.Sonar: self.__sonar})
-        self.__metric = self.class_under_test(subject=FakeSubject(self.__sonar), project=project)
+        self.__metric = self.class_under_test(subject=FakeSubject(), project=project)
 
     def test_value(self):
         """ Test that the value of the metric equals the value reported by Sonar. """
@@ -101,13 +100,13 @@ class CommonUnittestMetricTestsMixin(object):
 
     def test_is_applicable(self):
         """ Test that the metric is applicable. """
-        product = FakeSubject(self.__sonar)
+        product = FakeSubject()
         self.assertTrue(self.class_under_test.is_applicable(product))
 
     def test_is_not_applicable_with_integration_tests(self):
         """ Test that the metric isn't applicable when the product also has integration tests since then the combined
             unit and integration test coverage will be measured instead. """
-        product = FakeSubject(self.__sonar, integration_tests=True)
+        product = FakeSubject(integration_tests=True)
         self.assertFalse(self.class_under_test.is_applicable(product))