+++ docs/CHANGES.md
@@ -1,6 +1,6 @@
 2017-02-08  Release 1.77.0
 
-  * It's no longer necessary to specify unittests as a sub component.
+  * It's no longer necessary to specify unittests and integration tests as a sub component.
   
 
 2017-02-04  Release 1.76.4